+++ main.py
@@ -14,7 +14,7 @@ def validate_current_repository():
     exclude = os.getenv('INPUT_EXCLUDE', '')
 
     (status, successful, failed, ignored) = validate_cwd(exclude)
-    print(f'Exiting with status {status}, {len(successful)} successful, {len(failed)} failed, {len(ignored)} ignored.')
+    print(f'Exiting with status {status}: {len(successful)} successful, {len(failed)} failed, {len(ignored)} ignored.')
     exit(status)