+++ tests/singlefiles.py
@@ -1,8 +1,10 @@
+import json
 import os
 from pathlib import Path
 from unittest import TestCase
 
 import validator.validator as validator
+from validator.versionfile import VersionFile
 from .test_utils import schema, build_map
 
 
@@ -18,8 +20,11 @@ class TestSingleFiles(TestCase):
         os.chdir(cls.old_cwd)
 
     def test_invalidRemote(self):
-        (status, successful, failed, ignored) = validator.validate_cwd('', schema, build_map)
-        self.assertIn(Path('invalid-remote.version'), failed)
+        f = Path('./invalid-remote.version')
+        with f.open('r') as vf:
+            version_file = VersionFile(vf.read())
+            with self.assertRaises(json.decoder.JSONDecodeError):
+                remote = version_file.get_remote()
 
     def test_validRemote(self):
         (status, successful, failed, ignored) = validator.validate_cwd('', schema, build_map)